Updated stuff to 1.15.1 // Fixed wall closet

+++ src/main/java/tk/themcbros/uselessmod/entity/UselessCowEntity.java
@@ -55,7 +55,7 @@ public class UselessCowEntity extends AnimalEntity {
 	protected void registerAttributes() {
 		super.registerAttributes();
 		this.getAttribute(SharedMonsterAttributes.MAX_HEALTH).setBaseValue(10.0D);
-		this.getAttribute(SharedMonsterAttributes.MOVEMENT_SPEED).setBaseValue((double) 0.2F);
+		this.getAttribute(SharedMonsterAttributes.MOVEMENT_SPEED).setBaseValue(0.2D);
 	}
 
 	@Override
